New determineShelleyBasedEraWithEraInMode_ function
